pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codenames-benchmark"
version = "0.1.0"
description = "LLM benchmark suite where AI agents compete in Codenames"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "instructor>=1.0",
    "pydantic>=2.0",
    "typer>=0.15",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "tqdm>=4.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
codenames = "codenames.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/codenames"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
